requirements.txt
# AstroGuide Agent Dependencies

# Google Agent Development Kit
google-adk>=1.0.0

# Astronomical image analysis (for annotator.py)
astroquery>=0.4.6
astropy>=5.0
numpy>=1.21.0
Pillow>=9.0.0
python-dotenv>=1.0.0

